Enable multiple nodes sharing queue processor
This is a breaking change. Before the change, the queue processor
can only bind with a single node, so each node has to have at least
one dedicated thread to process the message. In this patch, a queue
processor can bind with multiple node, so the messages of different
nodes may be processed on the same thread, and that may saves resources
and reduce inter-node-communication latency.